move a few couch form utils to form_processor utils
