+++ corehq/apps/commtrack/parsing.py
@@ -59,7 +59,7 @@ from corehq.apps.commtrack import const
 from corehq.apps.commtrack.exceptions import InvalidDate
 from corehq.apps.commtrack.models import StockReportHelper, StockTransactionHelper
 from couchforms.models import XFormInstance
-from couchforms.util import adjust_datetimes
+from corehq.form_processor.utils import adjust_datetimes
 from xml2json.lib import convert_xml_to_json
