+++ corehq/ex-submodules/couchforms/__init__.py
@@ -1,3 +1,3 @@
 from .exceptions import XMLSyntaxError, CouchFormException
 from .getters import *
-from .util import SubmissionPost, fetch_and_wrap_form, convert_xform_to_json
+from .util import SubmissionPost, fetch_and_wrap_form
