+++ corehq/form_processor/tests/test_adjust_datetimes.py
@@ -2,7 +2,7 @@ from django.test import SimpleTestCase
 from corehq.apps.tzmigration import phone_timezones_should_be_processed
 from corehq.apps.tzmigration.test_utils import \
     run_pre_and_post_timezone_migration
-from couchforms.util import adjust_datetimes
+from corehq.form_processor.utils import adjust_datetimes
 
 
 class AdjustDatetimesTest(SimpleTestCase):
