+++ custom/uth/utils.py
@@ -1,6 +1,6 @@
 from casexml.apps.case.xform import get_case_updates
 from corehq.apps.receiverwrapper import submit_form_locally
-from couchforms import convert_xform_to_json
+from corehq.form_processor.utils import convert_xform_to_json
 from dimagi.utils.couch.database import get_db
 from casexml.apps.case.models import CommCareCase
 from lxml import etree
